% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clocks.R
\name{bday_clock}
\alias{bday_clock}
\title{Birthday clock}
\usage{
bday_clock(month, day)
}
\arguments{
\item{month}{English name for the month of your birth; see \code{\link[base]{month.name}}}

\item{day}{integer day of the month in which you were born; should be one of \emph{1-31}}
}
\description{
Birthday clock
}
\examples{
bday_clock("January", 1)
}
