% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{futurepop}
\alias{futurepop}
\title{Future population projection data}
\format{a time series object with 25 observations}
\usage{
futurepop
}
\description{
This dataset combines \code{\link[datasets]{uspop}} with some totally contrived, bogus, fictional population projects. The data starts in 1790 and goes through 2030 at 10 year increments.
}
\examples{
futurepop
plot(futurepop)
str(futurepop)
}
\keyword{datasets}
