% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(name = "user")
}
\arguments{
\item{name}{specify what you would like the program to call you;}
}
\description{
Say hello
}
\examples{
hello()
hello("hal")
}
